+++ test/rocprim/test_warp_scan.hpp
@@ -1,6 +1,6 @@
 // MIT License
 //
-// Copyright (c) 2017-2024 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2025 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -133,6 +133,127 @@ typed_test_def(RocprimWarpScanTests, name_suffix, InclusiveScan)
 
 }
 
+typed_test_def(RocprimWarpScanTests, name_suffix, InclusiveScanInitialValue)
+{
+    int device_id = test_common_utils::obtain_device_from_ctest();
+    SCOPED_TRACE(testing::Message() << "with device_id = " << device_id);
+    HIP_CHECK(hipSetDevice(device_id));
+
+    using T = typename TestFixture::params::type;
+    // for bfloat16 and half we use double for host-side accumulation
+    using binary_op_type_host = typename test_utils::select_plus_operator_host<T>::type;
+    binary_op_type_host binary_op_host;
+    using acc_type = typename test_utils::select_plus_operator_host<T>::acc_type;
+
+    // logical warp side for warp primitive, execution warp size is always rocprim::warp_size()
+    static constexpr size_t logical_warp_size = TestFixture::params::warp_size;
+
+    // The different warp sizes
+    static constexpr size_t ws32 = size_t(ROCPRIM_WARP_SIZE_32);
+    static constexpr size_t ws64 = size_t(ROCPRIM_WARP_SIZE_64);
+
+    // Block size of warp size 32
+    static constexpr size_t block_size_ws32
+        = rocprim::detail::is_power_of_two(logical_warp_size)
+              ? rocprim::max<size_t>(ws32, logical_warp_size * 4)
+              : rocprim::max<size_t>((ws32 / logical_warp_size), 1) * logical_warp_size;
+
+    // Block size of warp size 64
+    static constexpr size_t block_size_ws64
+        = rocprim::detail::is_power_of_two(logical_warp_size)
+              ? rocprim::max<size_t>(ws64, logical_warp_size * 4)
+              : rocprim::max<size_t>((ws64 / logical_warp_size), 1) * logical_warp_size;
+
+    unsigned int current_device_warp_size;
+    HIP_CHECK(::rocprim::host_warp_size(device_id, current_device_warp_size));
+
+    const size_t block_size = current_device_warp_size == ws32 ? block_size_ws32 : block_size_ws64;
+    const unsigned int grid_size = 4;
+    const size_t       size      = block_size * grid_size;
+
+    // Check if warp size is supported
+    if((logical_warp_size > current_device_warp_size)
+       || (current_device_warp_size != ws32
+           && current_device_warp_size != ws64)) // Only WarpSize 32 and 64 is supported
+    {
+        printf("Unsupported test warp size/computed block size: %zu/%zu. Current device warp size: "
+               "%u.    Skipping test\n",
+               logical_warp_size,
+               block_size,
+               current_device_warp_size);
+        GTEST_SKIP();
+    }
+
+    for(size_t seed_index = 0; seed_index < number_of_runs; seed_index++)
+    {
+        unsigned int seed_value
+            = seed_index < random_seeds_count ? rand() : seeds[seed_index - random_seeds_count];
+        SCOPED_TRACE(testing::Message() << "with seed = " << seed_value);
+
+        // Generate data
+        std::vector<T> input = test_utils::get_random_data<T>(size, 2, 50, seed_value);
+        std::vector<T> output(size);
+        std::vector<T> expected(output.size(), T(0));
+        T              initial_value = test_utils::get_random_data<T>(1, 2, 50, seed_value)[0];
+        SCOPED_TRACE(testing::Message() << "with initial_value = " << initial_value);
+
+        // Calculate expected results on host
+        for(size_t i = 0; i < input.size() / logical_warp_size; i++)
+        {
+            acc_type accumulator(initial_value);
+            for(size_t j = 0; j < logical_warp_size; j++)
+            {
+                auto idx      = i * logical_warp_size + j;
+                accumulator   = binary_op_host(input[idx], accumulator);
+                expected[idx] = static_cast<T>(accumulator);
+            }
+        }
+
+        // Writing to device memory
+        test_utils::device_ptr<T> device_input(input);
+        test_utils::device_ptr<T> device_output(output.size());
+
+        // Launching kernel
+        if(current_device_warp_size == ws32)
+        {
+            hipLaunchKernelGGL(
+                HIP_KERNEL_NAME(warp_inclusive_scan_initial_value_kernel<T,
+                                                                         block_size_ws32,
+                                                                         logical_warp_size>),
+                dim3(grid_size),
+                dim3(block_size),
+                0,
+                0,
+                device_input.get(),
+                device_output.get(),
+                initial_value);
+        }
+        else if(current_device_warp_size == ws64)
+        {
+            hipLaunchKernelGGL(
+                HIP_KERNEL_NAME(warp_inclusive_scan_initial_value_kernel<T,
+                                                                         block_size_ws64,
+                                                                         logical_warp_size>),
+                dim3(grid_size),
+                dim3(block_size),
+                0,
+                0,
+                device_input.get(),
+                device_output.get(),
+                initial_value);
+        }
+
+        HIP_CHECK(hipGetLastError());
+        HIP_CHECK(hipDeviceSynchronize());
+
+        // Read from device memory
+        output = device_output.load();
+
+        // Validating results
+        test_utils::assert_near(output, expected, test_utils::precision<T> * logical_warp_size);
+    }
+}
+
 typed_test_def(RocprimWarpScanTests, name_suffix, InclusiveScanReduce)
 {
     int device_id = test_common_utils::obtain_device_from_ctest();
@@ -252,6 +373,137 @@ typed_test_def(RocprimWarpScanTests, name_suffix, InclusiveScanReduce)
                                 test_utils::precision<T> * logical_warp_size);
     }
 
+}
+
+typed_test_def(RocprimWarpScanTests, name_suffix, InclusiveScanReduceInitialValue)
+{
+    int device_id = test_common_utils::obtain_device_from_ctest();
+    SCOPED_TRACE(testing::Message() << "with device_id = " << device_id);
+    HIP_CHECK(hipSetDevice(device_id));
+
+    using T = typename TestFixture::params::type;
+    // for bfloat16 and half we use double for host-side accumulation
+    using binary_op_type_host = typename test_utils::select_plus_operator_host<T>::type;
+    binary_op_type_host binary_op_host;
+    using acc_type = typename test_utils::select_plus_operator_host<T>::acc_type;
+
+    // logical warp side for warp primitive, execution warp size is always rocprim::warp_size()
+    static constexpr size_t logical_warp_size = TestFixture::params::warp_size;
+
+    // The different warp sizes
+    static constexpr size_t ws32 = size_t(ROCPRIM_WARP_SIZE_32);
+    static constexpr size_t ws64 = size_t(ROCPRIM_WARP_SIZE_64);
+
+    // Block size of warp size 32
+    static constexpr size_t block_size_ws32
+        = rocprim::detail::is_power_of_two(logical_warp_size)
+              ? rocprim::max<size_t>(ws32, logical_warp_size * 4)
+              : rocprim::max<size_t>((ws32 / logical_warp_size), 1) * logical_warp_size;
+
+    // Block size of warp size 64
+    static constexpr size_t block_size_ws64
+        = rocprim::detail::is_power_of_two(logical_warp_size)
+              ? rocprim::max<size_t>(ws64, logical_warp_size * 4)
+              : rocprim::max<size_t>((ws64 / logical_warp_size), 1) * logical_warp_size;
+
+    unsigned int current_device_warp_size;
+    HIP_CHECK(::rocprim::host_warp_size(device_id, current_device_warp_size));
+
+    const size_t block_size = current_device_warp_size == ws32 ? block_size_ws32 : block_size_ws64;
+    const unsigned int grid_size = 4;
+    const size_t       size      = block_size * grid_size;
+
+    // Check if warp size is supported
+    if((logical_warp_size > current_device_warp_size)
+       || (current_device_warp_size != ws32
+           && current_device_warp_size != ws64)) // Only WarpSize 32 and 64 is supported
+    {
+        printf("Unsupported test warp size/computed block size: %zu/%zu. Current device warp size: "
+               "%u.    Skipping test\n",
+               logical_warp_size,
+               block_size,
+               current_device_warp_size);
+        GTEST_SKIP();
+    }
+
+    for(size_t seed_index = 0; seed_index < number_of_runs; seed_index++)
+    {
+        unsigned int seed_value
+            = seed_index < random_seeds_count ? rand() : seeds[seed_index - random_seeds_count];
+        SCOPED_TRACE(testing::Message() << "with seed = " << seed_value);
+
+        // Generate data
+        std::vector<T> input = test_utils::get_random_data<T>(size, 2, 50, seed_value);
+        std::vector<T> output(size);
+        std::vector<T> output_reductions(size / logical_warp_size);
+        std::vector<T> expected(output.size(), T(0));
+        std::vector<T> expected_reductions(output_reductions.size(), T(0));
+        T              initial_value = test_utils::get_random_data<T>(1, 2, 50, seed_value)[0];
+        SCOPED_TRACE(testing::Message() << "with initial_value = " << initial_value);
+
+        // Calculate expected results on host
+        for(size_t i = 0; i < output.size() / logical_warp_size; i++)
+        {
+            acc_type accumulator(initial_value);
+            for(size_t j = 0; j < logical_warp_size; j++)
+            {
+                auto idx      = i * logical_warp_size + j;
+                accumulator   = binary_op_host(input[idx], accumulator);
+                expected[idx] = static_cast<T>(accumulator);
+            }
+            expected_reductions[i] = expected[(i + 1) * logical_warp_size - 1];
+        }
+
+        // Writing to device memory
+        test_utils::device_ptr<T> device_input(input);
+        test_utils::device_ptr<T> device_output(output.size());
+        test_utils::device_ptr<T> device_output_reductions(output_reductions.size());
+
+        // Launching kernel
+        if(current_device_warp_size == ws32)
+        {
+            hipLaunchKernelGGL(
+                HIP_KERNEL_NAME(warp_inclusive_scan_reduce_initial_value_kernel<T,
+                                                                                block_size_ws32,
+                                                                                logical_warp_size>),
+                dim3(grid_size),
+                dim3(block_size_ws32),
+                0,
+                0,
+                device_input.get(),
+                device_output.get(),
+                device_output_reductions.get(),
+                initial_value);
+        }
+        else if(current_device_warp_size == ws64)
+        {
+            hipLaunchKernelGGL(
+                HIP_KERNEL_NAME(warp_inclusive_scan_reduce_initial_value_kernel<T,
+                                                                                block_size_ws64,
+                                                                                logical_warp_size>),
+                dim3(grid_size),
+                dim3(block_size_ws64),
+                0,
+                0,
+                device_input.get(),
+                device_output.get(),
+                device_output_reductions.get(),
+                initial_value);
+        }
+
+        HIP_CHECK(hipGetLastError());
+        HIP_CHECK(hipDeviceSynchronize());
+
+        // Read from device memory
+        output            = device_output.load();
+        output_reductions = device_output_reductions.load();
+
+        // Validating results
+        test_utils::assert_near(output, expected, test_utils::precision<T> * logical_warp_size);
+        test_utils::assert_near(output_reductions,
+                                expected_reductions,
+                                test_utils::precision<T> * logical_warp_size);
+    }
 }
 
 typed_test_def(RocprimWarpScanTests, name_suffix, ExclusiveScan)